Working with out of index fanout values
